Added fmt::Debug for aiocb ReadAt and WriteAt for Ilumos
Although compio compiles and runs great on Illumos, it was only possible after resolving a rust print debug issue for aiocb.

In the solution, we add fmt::Debug for both ReadAt and WriteAt.

This probably isn't the best long term solution, but it worked in a pinch.

Note, I had an unexplained difference between my recent github pull, and this up to date fork. The important changes are obvious even if some line numbers have changed to protect the innocent.